delete card check and conflict error

+++ controllers/CardsController.js
@@ -1,5 +1,6 @@
 import CardsRepository from '../repositories/CardsRepository.js';
-import { CARD_NOT_FOUND, CREATED_CODE } from '../utils/constants.js';
+import { CARD_NOT_FOUND, CREATED_CODE, NO_RIGHTS } from '../utils/constants.js';
+import ForbiddenError from '../utils/errors/ForbiddenError.js';
 import NotFoundError from '../utils/errors/NotFoundError.js';
 
 async function createCard(request, response, next) {
@@ -22,10 +23,14 @@ async function getAllCards(request, response, next) {
 
 async function deleteCard(request, response, next) {
   try {
-    const card = await CardsRepository.deleteById(request.params.id);
+    const card = await CardsRepository.getById(request.params.id);
     if (card === null) {
       throw new NotFoundError(CARD_NOT_FOUND);
     }
+    if (card.owner.id !== request.user._id) {
+      throw new ForbiddenError(NO_RIGHTS);
+    }
+    await CardsRepository.deleteById(request.params.id);
     response.send(card);
   } catch (error) {
     next(error);
